refactored the rest of the labels
