+++ assets/styles/Labels.less
@@ -24,10 +24,9 @@
 }
 
 .Label__icon {
+    display: flex;
     position: relative;
-    width: 16px;
-    height: 16px;
-    margin-right: 3px;
+    margin-right: 4px;
 }
 
 .Label__discriminator {
